support linkstate weights in peers network
