+++ zenoh/src/net/protocol/mod.rs
@@ -12,3 +12,7 @@
 //   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
 //
 pub(crate) mod linkstate;
+pub(crate) mod network;
+
+pub(crate) const ROUTERS_NET_NAME: &str = "[Routers Network]";
+pub(crate) const PEERS_NET_NAME: &str = "[Peers Network]";
