+++ zenoh/src/net/routing/hat/linkstate_peer/pubsub.rs
@@ -31,22 +31,25 @@ use zenoh_protocol::{
 use zenoh_sync::get_mut_unchecked;
 
 use super::{
-    face_hat, face_hat_mut, get_peer, hat, hat_mut, network::Network, push_declaration_profile,
-    res_hat, res_hat_mut, HatCode, HatContext, HatFace, HatTables,
+    face_hat, face_hat_mut, get_peer, hat, hat_mut, push_declaration_profile, res_hat, res_hat_mut,
+    HatCode, HatContext, HatFace, HatTables,
 };
 #[cfg(feature = "unstable")]
 use crate::key_expr::KeyExpr;
-use crate::net::routing::{
-    dispatcher::{
-        face::FaceState,
-        interests::RemoteInterest,
-        pubsub::SubscriberInfo,
-        resource::{NodeId, Resource, SessionContext},
-        tables::{Route, RoutingExpr, Tables},
+use crate::net::{
+    protocol::network::Network,
+    routing::{
+        dispatcher::{
+            face::FaceState,
+            interests::RemoteInterest,
+            pubsub::SubscriberInfo,
+            resource::{NodeId, Resource, SessionContext},
+            tables::{Route, RoutingExpr, Tables},
+        },
+        hat::{CurrentFutureTrait, HatPubSubTrait, SendDeclare, Sources},
+        router::disable_matches_data_routes,
+        RoutingContext,
     },
-    hat::{CurrentFutureTrait, HatPubSubTrait, SendDeclare, Sources},
-    router::disable_matches_data_routes,
-    RoutingContext,
 };
 
 #[inline]
