+++ zenoh/src/net/routing/hat/linkstate_peer/token.rs
@@ -27,14 +27,17 @@ use zenoh_protocol::{
 use zenoh_sync::get_mut_unchecked;
 
 use super::{
-    face_hat, face_hat_mut, get_peer, hat, hat_mut, network::Network, push_declaration_profile,
-    res_hat, res_hat_mut, HatCode, HatContext, HatFace, HatTables,
+    face_hat, face_hat_mut, get_peer, hat, hat_mut, push_declaration_profile, res_hat, res_hat_mut,
+    HatCode, HatContext, HatFace, HatTables,
 };
-use crate::net::routing::{
-    dispatcher::{face::FaceState, interests::RemoteInterest, tables::Tables},
-    hat::{CurrentFutureTrait, HatTokenTrait, SendDeclare},
-    router::{NodeId, Resource, SessionContext},
-    RoutingContext,
+use crate::net::{
+    protocol::network::Network,
+    routing::{
+        dispatcher::{face::FaceState, interests::RemoteInterest, tables::Tables},
+        hat::{CurrentFutureTrait, HatTokenTrait, SendDeclare},
+        router::{NodeId, Resource, SessionContext},
+        RoutingContext,
+    },
 };
 
 #[inline]
