+++ zenoh/src/net/routing/hat/router/queries.rs
@@ -37,13 +37,13 @@ use zenoh_protocol::{
 use zenoh_sync::get_mut_unchecked;
 
 use super::{
-    face_hat, face_hat_mut, get_peer, get_router, hat, hat_mut, network::Network,
-    push_declaration_profile, res_hat, res_hat_mut, HatCode, HatContext, HatFace, HatTables,
+    face_hat, face_hat_mut, get_peer, get_router, hat, hat_mut, push_declaration_profile, res_hat,
+    res_hat_mut, HatCode, HatContext, HatFace, HatTables,
 };
 #[cfg(feature = "unstable")]
 use crate::key_expr::KeyExpr;
 use crate::net::{
-    protocol::linkstate::LinkEdge,
+    protocol::{linkstate::LinkEdge, network::Network},
     routing::{
         dispatcher::{
             face::FaceState,
