+++ zenoh/src/net/routing/hat/router/token.rs
@@ -27,11 +27,11 @@ use zenoh_protocol::{
 use zenoh_sync::get_mut_unchecked;
 
 use super::{
-    face_hat, face_hat_mut, get_peer, get_router, hat, hat_mut, network::Network,
-    push_declaration_profile, res_hat, res_hat_mut, HatCode, HatContext, HatFace, HatTables,
+    face_hat, face_hat_mut, get_peer, get_router, hat, hat_mut, push_declaration_profile, res_hat,
+    res_hat_mut, HatCode, HatContext, HatFace, HatTables,
 };
 use crate::net::{
-    protocol::linkstate::LinkEdge,
+    protocol::{linkstate::LinkEdge, network::Network},
     routing::{
         dispatcher::{face::FaceState, interests::RemoteInterest, tables::Tables},
         hat::{CurrentFutureTrait, HatTokenTrait, SendDeclare},
